Set convolution-based models to be on bfp16 (#2067)
### Ticket

- https://github.com/tenstorrent/tt-forge-fe/issues/1987

### Summary 

- Set convolution-based models to be on bfp16 

### Note 

- New regressions are tracked in
[#1987](https://github.com/tenstorrent/tt-forge-fe/issues/1987).

- In the VGG models, the PCC for vgg19 and bn_vgg19b is below 0.99 even
in the float32 case. As [Everything above 0.95 is tolerable for
now](https://github.com/tenstorrent/tt-forge-fe/issues/1987#issuecomment-2891165285)
pcc adjusted accordingly.

| Test Name       | PCC (float32) | PCC (bfloat16) |
|----------------|----------------|-----------------|
|
forge/test/models/pytorch/vision/vgg/test_vgg.py::test_vgg_osmr_pytorch[vgg16]
| >=0.99 | 0.9895039351320837 |
|
forge/test/models/pytorch/vision/vgg/test_vgg.py::test_vgg_osmr_pytorch[vgg19]
| 0.9856318375050916 | 0.9772661805379736 |
|
forge/test/models/pytorch/vision/vgg/test_vgg.py::test_vgg_osmr_pytorch[bn_vgg19b]
| 0.986650973327936 | 0.9691731206520947 |

+++ forge/test/models/pytorch/vision/mnist/model_utils/utils.py
@@ -1,6 +1,7 @@
 # SPDX-FileCopyrightText: (c) 2025 Tenstorrent AI ULC
 #
 # SPDX-License-Identifier: Apache-2.0
+import torch
 from torch.utils.data import DataLoader
 from torchvision import datasets, transforms
 
@@ -18,4 +19,4 @@ def load_input():
     test_dataset = datasets.MNIST(root="./data", train=False, transform=transform, download=True)
     dataloader = DataLoader(test_dataset, batch_size=1)
     test_input, _ = next(iter(dataloader))
-    return [test_input]
+    return [test_input.to(torch.bfloat16)]
