+++ forge/test/models/pytorch/vision/regnet/model_utils/image_utils.py
@@ -3,6 +3,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import requests
+import torch
 from PIL import Image
 from transformers import AutoImageProcessor
 
@@ -19,4 +20,4 @@ def preprocess_input_data(image_url, variant):
     # Preprocess the image
     image_tensor = preprocessor(images=image, return_tensors="pt").pixel_values
 
-    return [image_tensor]
+    return [image_tensor.to(torch.bfloat16)]
